Move examples to test source set
